+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 1.0.0 (unreleased)
 ------------------
 
+- #11 Additional `send` transition in AR/Sample workflow
 - #7 New field "Study code" in Client (instead of Tax Number)
 - #2 New fields in Analysis Request (and visible in AR Add)
 - #1 Hide unused fields from Analysis Request Add form